Move load_json to misc
